requirements.txt
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.4
django-rest-swagger==2.2.0
djangorestframework==3.10.2
Faker==0.9.1
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
simplejson==3.16.0
sqlparse==0.3.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.25.3
